Replace half with float

+++ colbert/indexing/codecs/residual.py
@@ -24,7 +24,7 @@ class ResidualCodec:
         ResidualCodec.try_load_torch_extensions(self.use_gpu)
 
         if self.use_gpu > 0:
-            self.centroids = centroids.cuda().half()
+            self.centroids = centroids.cuda().float()
         else:
             self.centroids = centroids.float()
         self.dim, self.nbits = config.dim, config.nbits
@@ -32,12 +32,12 @@ class ResidualCodec:
 
         if torch.is_tensor(self.avg_residual):
             if self.use_gpu:
-                self.avg_residual = self.avg_residual.cuda().half()
+                self.avg_residual = self.avg_residual.cuda().float()
 
         if torch.is_tensor(bucket_cutoffs):
             if self.use_gpu:
                 bucket_cutoffs = bucket_cutoffs.cuda()
-                bucket_weights = bucket_weights.half().cuda()
+                bucket_weights = bucket_weights.float().cuda()
 
         self.bucket_cutoffs = bucket_cutoffs
         self.bucket_weights = bucket_weights
@@ -169,7 +169,7 @@ class ResidualCodec:
 
         for batch in embs.split(1 << 18):
             if self.use_gpu:
-                batch = batch.cuda().half()
+                batch = batch.cuda().float()
             codes_ = self.compress_into_codes(batch, out_device=batch.device)
             centroids_ = self.lookup_centroids(codes_, out_device=batch.device)
 
@@ -212,7 +212,7 @@ class ResidualCodec:
         bsize = (1 << 29) // self.centroids.size(0)
         for batch in embs.split(bsize):
             if self.use_gpu:
-                indices = (self.centroids @ batch.T.cuda().half()).max(dim=0).indices.to(device=out_device)
+                indices = (self.centroids @ batch.T.cuda().float()).max(dim=0).indices.to(device=out_device)
             else:
                 indices = (self.centroids @ batch.T.cpu().float()).max(dim=0).indices.to(device=out_device)
             codes.append(indices)
@@ -268,7 +268,7 @@ class ResidualCodec:
                 centroids_.add_(residuals_)
 
             if self.use_gpu:
-                D_ = torch.nn.functional.normalize(centroids_, p=2, dim=-1).half()
+                D_ = torch.nn.functional.normalize(centroids_, p=2, dim=-1).float()
             else:
                 D_ = torch.nn.functional.normalize(centroids_.to(torch.float32), p=2, dim=-1)
             D.append(D_)
